Auth n auth z

+++ controllers/user_controller.js
@@ -2,6 +2,7 @@ import Individual_User from "../models/user_models/individual_user_model.js";
 import Company_User from "../models/user_models/company_user_model.js";
 import Admin_User from "../models/user_models/admin_user_model.js";
 import Common_User from "../models/user_models/common_user_model.js";
+import * as Pure_OTP from "../utils/pure_otp.js";
 import Auth from "../models/auth_model.js";
 import jwt from "jsonwebtoken";
 import send_email from "../utils/send_email.js";
@@ -58,31 +59,37 @@ const register = async (_req, _res) => {
 		const user_register_info = {..._req.body};
 		delete user_register_info.role;
 
-		const user_register_info_password_bcrypt = await bcrypt.hash(_req.body.password, 10);
+		const user_register_info_password_bcrypt = await bcrypt.hash(Buffer.from(_req.body.password, "utf-8"), 10);
 		user_register_info.password = user_register_info_password_bcrypt;
 		//----- User Register to DB & Response
 		try {
-			//TODO: Mail gitmeyebilir tekrar al butonu atmamız gerekicek.
 			const user_registered = await DB_access.create(user_register_info);
 			if (!user_registered) {
 				throw new Error("User Register Error");
 			}
 			const userId = user_registered._id.toString();
-			await Common_User.create({email: _req.body.email, role: _req.body.role , user_id: userId});
+			await Common_User.create({email: _req.body.email, role: _req.body.role, user_id: userId});
 			//------------------
 			const query = {};
 			query[user_id_field] = user_registered._id;
-			const confirm_credential = crypto.randomInt(100000, 999999).toString();
-			const verification_code = await Verification.create({...query, verification_code: confirm_credential});
-
+			const secret_key = Pure_OTP.generateSecretKey();
+			const confirm_credential = Pure_OTP.generateOTP(secret_key);
+			const verification_code = await Verification.create({...query, verification_code: confirm_credential, created_at: Date.now()});
 			if (!verification_code) throw new Error("Verification Code Error");
 			//------------------
-			let confirm_url_token = jwt.sign({user_id: user_registered._id, role: _req.body.role}, process.env.CONFIRM_TOKEN_SECRET, {expiresIn: "30m"});
-			confirm_url_token = btoa(confirm_url_token);
+			const confirm_token = jwt.sign({user_id: user_registered._id, role: _req.body.role}, process.env.CONFIRM_TOKEN_SECRET, {expiresIn: "1h"});
 			send_email(_req.body.email, "Confirm account 🤕", "confirm_account", confirm_credential);
-			//------------------
+
+			_res.cookie("confirm_token", confirm_token, {
+				maxAge: 60 * 60 * 1000,
+				sameSite: "None",
+				// domain: "localhost",
+				secure: true,
+				httpOnly: true,
+			}); //------------------
+
 			console.info(chalk.green.bold(`${getTimestamp()} Status Code : 201 -- Info : User Created -- ID : ${user_registered._id}`));
-			return _res.status(201).json({message: "User Created , Please Confirm Your Email", confirm_url_token, status_code: "201", status: "success", time: "5"});
+			return _res.status(201).json({message: "User Created , Please Confirm Your Email", status_code: "201", status: "success"});
 		} catch (error) {
 			console.error(chalk.bold(`${getTimestamp()} Status Code : 400 -- Error : ${error} -- Service : Register`));
 			return _res.status(400).json({message: "Invalid User Data", status_code: "400", status: "error"});
@@ -101,19 +108,12 @@ const confirm = async (_req, _res) => {
 		let DB_access = "";
 		let user_id_field = "";
 		let confirm_user = "";
-		let authHeader = _req.headers.Authorization || _req.headers.authorization;
-		if (!authHeader || !authHeader.startsWith("Bearer")) {
-			console.error(chalk.bold(`${getTimestamp()} Status Code : 400 -- Error : No confirm token -- Service : confirm_access_token`));
-			return _res.status(400).json({message: "Not confirm , No confirm token", status_code: "400", status: "error"});
-		}
-
-		// Extract the token from the header
-		let token = authHeader.split(" ")[1];
-		token = atob(token);
+		const token = _req.cookies.confirm_token;
+		const requestReceivedTime = Date.now();
 		// Verify the token
 		jwt.verify(token, process.env.CONFIRM_TOKEN_SECRET, (err, decoded) => {
 			if (err) {
-				console.error(chalk.bold(`${getTimestamp()} Status Code : 403 -- Error : ${err} -- Service : confirm_access_token`));
+				console.error(chalk.bold(`${getTimestamp()} Status Code : 403 -- Error : ${err} -- Service : confirm_token`));
 				return _res.status(403).json({message: "Not authorized, no valid token", status_code: "403", status: "error"});
 			}
 			// If the token is valid, set the user in the request and proceed to the next middleware
@@ -132,17 +132,20 @@ const confirm = async (_req, _res) => {
 		//----- DB Access Type Check
 		if (confirm_user.role === "Company_User") DB_access = Company_User;
 		if (confirm_user.role === "Individual_User") DB_access = Individual_User;
-		if (confirm_user.role === "Admin_User") DB_access = Admin_User;		
+		if (confirm_user.role === "Admin_User") DB_access = Admin_User;
 
 		try {
 			const query = {};
 			query[user_id_field] = confirm_user.user_id;
 			const validate_user = await Verification.findOne({...query, verification_code: _req.body.confirm_credential});
 			if (!validate_user) throw new Error("Verification Code DB Search Error");
+			if (requestReceivedTime - validate_user.created_at > 60000) throw new Error("Verification Code Expired");
 			if (validate_user.verification_code !== _req.body.confirm_credential) throw new Error("Verification Code Not Match");
 
 			await Verification.deleteOne({...query});
-			await DB_access.updateOne({_id: confirm_user.user_id }, {status: "ACTIVE"});
+			await DB_access.updateOne({_id: confirm_user.user_id}, {status: "ACTIVE"});
+
+			_res.clearCookie("confirm_token");
 
 			console.info(chalk.green.bold(`${getTimestamp()} Status Code : 200 -- Info : User Verified -- ID : ${confirm_user.user_id}`));
 			return _res.status(200).json({message: "User Verified", status_code: "200", status: "success"});
@@ -163,17 +166,8 @@ const re_confirm = async (_req, _res) => {
 	try {
 		let user_id_field = "";
 		let confirm_user = "";
-		let authHeader = _req.headers.Authorization || _req.headers.authorization;
 		let DB_access = "";
-		if (!authHeader || !authHeader.startsWith("Bearer")) {
-			console.error(chalk.bold(`${getTimestamp()} Status Code : 400 -- Error : No confirm token -- Service : confirm_access_token`));
-			return _res.status(400).json({message: "Not confirm , No confirm token", status_code: "400", status: "error"});
-		}
-
-		// Extract the token from the header
-		let token = authHeader.split(" ")[1];
-		token = atob(token);
-
+		const token = _req.cookies.confirm_token;
 		// Verify the token
 		jwt.verify(token, process.env.CONFIRM_TOKEN_SECRET, (err, decoded) => {
 			if (err) {
@@ -181,9 +175,8 @@ const re_confirm = async (_req, _res) => {
 				return _res.status(403).json({message: "Not authorized, no valid token", status_code: "403", status: "error"});
 			}
 			// If the token is valid, set the user in the request and proceed to the next middleware
-			confirm_user = {user_id: decoded.user_id, role: decoded.role };
+			confirm_user = {user_id: decoded.user_id, role: decoded.role};
 		});
-
 		if (!(confirm_user.role !== "Company_User" || confirm_user.role !== "Individual_User" || confirm_user.role !== "Admin_User")) {
 			console.error(chalk.bold(`${getTimestamp()} Status Code : 400 -- Error : User Role Not Found -- Service : Re-Validate`));
 			return _res.status(400).send({message: "User Profile Not Found", status_code: "400", status: "error"});
@@ -201,17 +194,25 @@ const re_confirm = async (_req, _res) => {
 			const validate_user = await Verification.findOne({...query});
 			if (!validate_user) throw new Error("Verification Code DB Search Error");
 			await Verification.deleteOne({...query});
-			const confirm_credential = crypto.randomInt(100000, 999999).toString();
-			const verification_code = await Verification.create({...query, verification_code: confirm_credential});
+			const secret_key = Pure_OTP.generateSecretKey();
+			const confirm_credential = Pure_OTP.generateOTP(secret_key);
+			const verification_code = await Verification.create({...query, verification_code: confirm_credential, created_at: Date.now()});
 			if (!verification_code) throw new Error("Verification Code Creation Error");
 			//------------------
-			let confirm_url_token = jwt.sign({_id: confirm_user.user_id, role: confirm_user.role}, process.env.CONFIRM_TOKEN_SECRET, {expiresIn: "1h"}); //TODO: Can change , WE CAN NOT GIVE A NEW TOKEN BECAUSE ALREADY HAS TIME
-			confirm_url_token = btoa(confirm_url_token);
+			let confirm_token = jwt.sign({user_id: confirm_user.user_id, role: confirm_user.role}, process.env.CONFIRM_TOKEN_SECRET, {expiresIn: "1h"}); //TODO: Can change , WE CAN NOT GIVE A NEW TOKEN BECAUSE ALREADY HAS TIME
 
+			_res.cookie("confirm_token", confirm_token, {
+				maxAge: 60 * 60 * 1000,
+				sameSite: "None",
+				// domain: "localhost",
+				secure: true,
+				httpOnly: true,
+			});
+			//------------------
 			send_email(re_confirmed_user.email, "Confirm account 🤕", "confirm_account", confirm_credential);
 			//TODO: CAN CHANGE BECAUSE OF URL WE CAN GİVE AN REFRESH TOKEN ALSO FOR CONFIRM STEP AND THEN DELETE THEM. BUT NOW WE USE THIS
 			console.info(chalk.green.bold(`${getTimestamp()} Status Code : 200 -- Info : User Re Confirmed -- ID : ${confirm_user.user_id}`));
-			return _res.status(200).json({message: "Validate Token , resend please confirm", confirm_url_token, status_code: "200", status: "success"});
+			return _res.status(200).json({message: "Validate Token , resend please confirm", status_code: "200", status: "success"});
 		} catch (error) {
 			console.error(chalk.bold(`${getTimestamp()} Status Code : 400 -- Error : ${error} -- Service : Re-Validate`));
 			return _res.status(400).json({message: "User Verification Error", status_code: "400", status: "error"});
@@ -254,22 +255,40 @@ const login = async (_req, _res) => {
 
 		//----- User Check
 		try {
-			is_user_available = await DB_access.findOne({email: _req.body.email});
-			if (!is_user_available) throw new Error("User Not Found In Related DB");
+			const check_user_availability = await DB_access.findOne({email: _req.body.email});
+			if (!check_user_availability) throw new Error("User Not Found In Related DB");
+			is_user_available = {is_user_available, check_user_availability};
 		} catch (error) {
 			console.error(chalk.bold(`${getTimestamp()} Status Code : 401 -- Error : ${error} -- Service : Login`));
 			return _res.status(401).json({message: "Invalid Credentials", status_code: "401", status: "error"});
 		}
 		//----- User Status Check
-		if (is_user_available.status !== "ACTIVE") {
-			console.error(chalk.bold(`${getTimestamp()} Status Code : 401 -- Error : User is not active -- Service : Login -- ID : ${is_user_available.email}`));
+		if (is_user_available.check_user_availability.status !== "ACTIVE") {
+			const query = {};
+			query[user_id_field] = is_user_available.check_user_availability._id;
+			const secret_key = Pure_OTP.generateSecretKey();
+			const confirm_credential = Pure_OTP.generateOTP(secret_key);
+			const verification_code = await Verification.create({...query, verification_code: confirm_credential, created_at: Date.now()});
+			if (!verification_code) throw new Error("Verification Code Error");
+			//------------------
+			const confirm_token = jwt.sign({user_id: is_user_available.check_user_availability._id, role: is_user_available.is_user_available.role}, process.env.CONFIRM_TOKEN_SECRET, {expiresIn: "1h"});
+			send_email(_req.body.email, "Confirm account 🤕", "confirm_account", confirm_credential);
+
+			_res.cookie("confirm_token", confirm_token, {
+				maxAge: 60 * 60 * 1000,
+				sameSite: "None",
+				// domain: "localhost",
+				secure: true,
+				httpOnly: true,
+			}); //------------------
+
+			console.error(chalk.bold(`${getTimestamp()} Status Code : 401 -- Error : User is not active -- Service : Login -- ID : ${is_user_available.is_user_available.email}`));
 			return _res.status(401).json({message: "User is not active", status_code: "401", status: "error"});
 		}
-		//TODO : Bu aşamada kullanıcı doğrulama ekranına gidecek ve tekrar bir doğrulama kodu isteyecek ardından doğrulama kodu doğruysa login işlemi gerçekleşecek.
 		//----- DB Access Type Check
 		try {
 			const query = {};
-			query[user_id_field] = is_user_available._id;
+			query[user_id_field] = is_user_available.check_user_availability._id;
 			const is_user_refresh_token_available = await Auth.findOne({...query});
 			if (is_user_refresh_token_available) throw new Error("User Already Logged In");
 		} catch (error) {
@@ -278,20 +297,20 @@ const login = async (_req, _res) => {
 		}
 		//----- Password Check
 		try {
-			is_password_match = await bcrypt.compare(_req.body.password, is_user_available.password);
+			is_password_match = await bcrypt.compare(Buffer.from(_req.body.password, "utf-8"), is_user_available.check_user_availability.password);
 			if (!is_password_match) throw new Error("Password Not Match");
 		} catch (error) {
 			console.error(chalk.bold(`${getTimestamp()} Status Code : 401 -- Error : ${error} -- Service : Login`));
 			return _res.status(401).json({message: "Invalid Credentials : Password", status_code: "401", status: "error"});
 		}
 		//----- JWT Token Create
-		const access_token = jwt.sign({_id: is_user_available._id, role: _req.body.role}, process.env.ACCESS_TOKEN_SECRET, {expiresIn: "1m"});
-		const refresh_token = jwt.sign({_id: is_user_available._id, role: _req.body.role}, process.env.REFRESH_TOKEN_SECRET, {expiresIn: "1d"});
+		const access_token = jwt.sign({_id: is_user_available.check_user_availability._id, role: is_user_available.is_user_available.role}, process.env.ACCESS_TOKEN_SECRET, {expiresIn: "1m"});
+		const refresh_token = jwt.sign({_id: is_user_available.check_user_availability._id, role: is_user_available.is_user_available.role}, process.env.REFRESH_TOKEN_SECRET, {expiresIn: "1d"});
 
 		//----- Token Save to DB & Cookie Set & Response
 		try {
 			const query = {};
-			query[user_id_field] = is_user_available._id;
+			query[user_id_field] = is_user_available.check_user_availability._id;
 			const current_user = {...query, refresh_token};
 
 			const auth_result = await Auth.create(current_user);
@@ -303,8 +322,8 @@ const login = async (_req, _res) => {
 				secure: true, // "true" yerine "true" olarak ayarlanmalı
 				httpOnly: true, // "true" yerine "true" olarak ayarlanmalı
 			});
-			console.info(chalk.green.bold(`${getTimestamp()} Status Code : 200 -- Info : User Authenticated -- ID : ${is_user_available._id}`));
-			return _res.status(200).json({message: "User Authenticated", access_token, status_code: "200", status: "success", name: is_user_available.name, surname: is_user_available.surname});
+			console.info(chalk.green.bold(`${getTimestamp()} Status Code : 200 -- Info : User Authenticated -- ID : ${is_user_available.check_user_availability._id}`));
+			return _res.status(200).json({message: "User Authenticated", access_token, status_code: "200", status: "success"});
 		} catch (error) {
 			console.error(chalk.bold(`${getTimestamp()} Status Code : 404 -- Error : ${error} -- Service : Login`));
 			return _res.status(404).json({message: "Invalid Auth Data", status_code: "404", status: "error"});
@@ -440,7 +459,7 @@ const reset_password = async (_req, _res) => {
 			return _res.status(400).json({message: "Invalid User Id"});
 		}
 		//-----
-		is_user_available.password = await bcrypt.hash(_req.body.password, 10);
+		is_user_available.password = await bcrypt.hash(Buffer.from(_req.body.password, "utf-8"), 10);
 		//-----
 		try {
 			const result = await is_user_available.save();
@@ -475,8 +494,15 @@ const current = async (_req, _res) => {
 		//-----
 		try {
 			const is_user_available = await DB_access.findOne({_id: _req.user._id});
+			if (!is_user_available) throw new Error("User Not Found");
+			const user_dto = {
+				_id: is_user_available._id,
+				name: is_user_available.name,
+				surname: is_user_available.surname,
+				email: is_user_available.email,
+			};
 			console.info(chalk.green.bold(`${getTimestamp()} Status Code : 200 -- Info : User Info Sent -- ID : ${is_user_available._id}`));
-			return _res.status(200).json(is_user_available);
+			return _res.status(200).json(user_dto);
 		} catch (error) {
 			console.error(chalk.bold(`${getTimestamp()} Status Code : 401 -- Error : ${error} -- Service : Current User`));
 			return _res.status(401).json({message: "Invalid Credentials"});
